Moving HtmlTextContainer to designsystem.
